병합 전 마지막 점검

+++ src/main/java/com/project/movie/domain/MovieDetailReplyDTO.java
@@ -15,7 +15,7 @@ public class MovieDetailReplyDTO {
 	private int rno;
 	private String replyContent;
 	private String userid;	
-	private float grade;
+	private int grade;
 	private Date replydate;	
 	private Date replyUpdatedate;
